Apply peter's java sp suggestions
Former-commit-id: 80e740890fe1842887ffb95c544498a6ebdd4766
Former-commit-id: b48519e8c8255dc0831e8994c635a858c654db66
Former-commit-id: 998922e9dc085d254c49d40d68cbaa49c5153606

+++ airesources/Java/Site.java
@@ -1,3 +1,3 @@
 public class Site{
-  public short owner, strength, production;
+  public int owner, strength, production;
 }
